refactor: create BattleCapture component
